Rename of models folder, setup of facade and services, and projects (infrastructure, application). Minimal interface implentation, no working code. Pages are unaffected as of now

+++ GymSwipe.Domain/Models/ExerciseRecords.cs
@@ -13,7 +13,7 @@ namespace GymSwipe.Domain.Domain
         public int ExerciseId { get; set; }
         public Exercise Exercise { get; set; } = null!;
         public int Repetitions { get; set; }
-        public int WeightKg { get; set; }
+        public double WeightKg { get; set; }
         public int Sets { get; set; }
         public DateTime DateRecorded { get; set; }
         public double UserRating { get; set; }
